agregue errorer a light sensors

+++ src/sensors/light_sensor.h
@@ -2,6 +2,6 @@
 #define LIGHT_SENSOR_H
 #include <stdint.h>
 void light_sensor_init(void);
-uint16_t light_sensor_read(void);
+int light_sensor_read(void);
 
 #endif
